+++ Sources/VirtualMeetingFinder/MainApp/VMF_AppDelegate.swift
@@ -1,5 +1,5 @@
 /*
- © Copyright 2024, Little Green Viper Software Development LLC
+ © Copyright 2024-2026, Little Green Viper Software Development LLC
  LICENSE:
  
  MIT License
@@ -90,7 +90,7 @@ extension VMF_AppDelegate {
      /**
       Fetches all of the virtual meetings (hybrid and pure virtual).
       
-      - parameter completion: A tail completion handler. One parameter, with the virtual service (nil, if error). May be called in any thread.
+      - parameter inCompletion: A tail completion handler. One parameter, with the virtual service (nil, if error). May be called in any thread.
       */
      class func findMeetings(completion inCompletion: @escaping (SwiftBMLSDK_MeetingLocalTimezoneCollection?) -> Void) {
           virtualService = nil
@@ -124,9 +124,9 @@ extension VMF_AppDelegate {
      /**
       Displays the given message and title in an alert with an "OK" button.
       
-      - parameter header: a string to be displayed as the title of the alert. It is localized by this method.
-      - parameter message: a string to be displayed as the message of the alert. It is localized by this method.
-      - parameter presentedBy: An optional UIViewController object that is acting as the presenter context for the alert. If nil, we use the top controller of the Navigation stack.
+      - parameter inHeader: a string to be displayed as the title of the alert. It is localized by this method.
+      - parameter inMessage: a string to be displayed as the message of the alert. It is localized by this method.
+      - parameter inPresentingViewController: An optional UIViewController object that is acting as the presenter context for the alert. If nil, we use the top controller of the Navigation stack.
       */
      class func displayAlert(header inHeader: String, message inMessage: String = "", presentedBy inPresentingViewController: UIViewController! = nil) {
           // This ensures that we are on the main thread.
@@ -171,9 +171,9 @@ extension VMF_AppDelegate: UIApplicationDelegate {
      /**
       This opens a URL.
       
-      - parameter url: The URL to open.
-      - parameter options: The URLoptions.
-      - parameter completionHandler: The closure to be executed, upon completion of the open. It has one parameter, a Boolean, that is true, if the open was successful. This is always called in the main thread.
+      - parameter inURL: The URL to open.
+      - parameter inOptions: The URLoptions.
+      - parameter inClosure: The closure to be executed, upon completion of the open. It has one parameter, a Boolean, that is true, if the open was successful. This is always called in the main thread.
       */
      class func open(url inURL: URL, options inOptions: [UIApplication.OpenExternalURLOptionsKey: Any] = [:], completionHandler inClosure: ((Bool) -> Void)? = nil) {
           UIApplication.shared.open(inURL, options: inOptions, completionHandler: inClosure)
@@ -183,22 +183,22 @@ extension VMF_AppDelegate: UIApplicationDelegate {
      /**
       Called when the app has set itself up, and is about to start.
       
-      - parameter: The application (ignored)
+      - parameter inApplication: The application (ignored)
       - parameter didFinishLaunchingWithOptions: Launch options (also ignored).
       - returns: True (always).
       */
-     func application(_: UIApplication, didFinishLaunchingWithOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool { true }
+     func application(_ inApplication: UIApplication, didFinishLaunchingWithOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool { true }
      
      /* ################################################################## */
      /**
       Called to deliver the scene configuration for the connection.
       
-      - parameter: The application (ignored)
-      - parameter configurationForConnecting: The session being connected.
+      - parameter inApplication: The application (ignored)
+      - parameter inConnectingSceneSession: The session being connected.
       - parameter options: Launch options (also ignored).
       - returns: A new instance of a default configuration for the scene.
       */
-     func application(_: UIApplication, configurationForConnecting inConnectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
+     func application(_ inApplication: UIApplication, configurationForConnecting inConnectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
           UISceneConfiguration(name: "Default Configuration", sessionRole: inConnectingSceneSession.role)
      }