+++ Sources/VirtualMeetingFinder/ViewControllers/Screens/SettingsScreen/VMF_SettingsViewController.swift
@@ -1,5 +1,5 @@
 /*
- © Copyright 2024, Little Green Viper Software Development LLC
+ © Copyright 2024-2026, Little Green Viper Software Development LLC
  LICENSE:
  
  MIT License